Arrivals: factor out an EtaStrip composable + scroll-affordance polish
Split the ETA strip out of the 1176-line ArrivalRows.kt into a self-contained
EtaStrip.kt: the strip (horizontal scroll + overflow chevrons + the
pull-past-the-end load-more gesture and reveal), the pills (EtaPill,
EtaPillWithMenu) and the per-trip menu (TripActionsMenu). RouteArrivalRow and
the legacy flat-row bits stay in ArrivalRows.kt and reach the shared helpers
(RealtimeIndicator, MenuRow, strikeThroughIf, the preview fixtures) cross-file;
external importers are unaffected (same components package).

Scroll-affordance polish while it was open:
- The "more to scroll" chevrons now sit OUTSIDE the scrollable content box, in
  reserved leading/trailing gutters (so their show/hide can't reflow the pills),
  with the edge fade at the content's matching edge — previously a single right
  chevron perched just inside, over the last pill.
- Added a symmetric LEFT chevron + left-edge fade shown once the strip is
  scrolled off its start (reuses the right chevron drawable rotated 180°).
  Factored the mirror pieces into ScrollChevronGutter and BoxScope.EdgeFade.
- The load-more pull chip draws its circle with Surface(shape = CircleShape).

Open the row with the soonest upcoming ETA first: EtaStrip takes a `start`
index and justifies that pill to the leading edge on first display (a one-shot,
scroll-independent measure-then-scrollTo), so recent-past pills overflow off the
LEFT — reachable via the new left chevron. RouteArrivalRow passes the boundary
the model already owns, RouteRowGroup.firstUpcomingIndex (first non-negative
ETA, the row's sort key).

Previews: EtaStrip fits / overflowing / scrolled (both chevrons) via a shared
EtaStripPreviewFrame, all bounded with height(IntrinsicSize.Min) so the
chevrons' fillMaxHeight resolves to the pill height; renamed the bare-pills
gallery to EtaPillVariantsPreview (it never used the strip). Scroll state is
hoisted as an optional EtaStrip param so a preview can seed a mid-scroll offset.
